requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
atomicwrites==1.4.0
attrs==20.3.0
more-itertools==8.7.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==5.1.1
pytest-rerunfailures==9.1.1
selenium==3.14.0
six==1.16.0
urllib3==1.26.4
wcwidth==0.2.5
